dabproject/pyproject.toml
[project]
name = "dabproject"
version = "0.0.1"
authors = [{ name = "gantagayatri2001@gmail.com" }]
requires-python = ">=3.10,<=3.13"

[dependency-groups]
dev = [
    "pytest",

    # Code completion support for Lakeflow Declarative Pipelines, also install databricks-connect
    "databricks-dlt",

    # databricks-connect can be used to run parts of this project locally.
    # Note that for local development, you should use a version that is not newer
    # than the remote cluster or serverless compute you connect to.
    # See also https://docs.databricks.com/dev-tools/databricks-connect.html.
    "databricks-connect>=15.4,<15.5",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dabproject"]

[project.scripts]
main = "dabproject.main:main"
